requirements.txt
requests>=2.31.0
pytest>=8.2.0
pytest-html>=4.1.1
pytest-metadata>=3.0.0
allure-pytest>=2.15.0
pyyaml>=6.0.1
